# optimize street.normalized_name = housenumber.normalized_street comparisons