@breunigs 2023-03-12
I am fairly certain there is a win to be had here, given that this query involves a sequential table scan. I assume at the moment that ~15s time saves on a ~15min job get lost in the noise when I use the computer (or it does backups or some such). It should be visible when either processing a larger area or improving the statistical reliability. Tackling the other larger improvements first made more sense, though.