@philippks 2023-03-12
Sounds reasonable. If we can verify this on larger areas, the change is very welcome 👍