@breunigs 2023-03-13
FWIW, this didn't help. The numbers for `set_street_ids_by_street_name.sql` are:
```
                  master: 23714s
optimize-wikipedia-index: 24031s (the one where the backup ran)
optimize-set-street-ids…: 25810s (no backup)
```
so this PR would actually mean a significant slowdown of 30min. Surprising!